BackEnd/requirements.txt
fastapi
uvicorn[standard]
langchain
openai
python-dotenv
PyPDF2
pypdf
chromadb
requests
python-multipart
langchain-openai 
pdfminer.six